server-py/pyproject.toml
[project]
name = "worldmonitor-server"
version = "2.0.0"
description = "WorldMonitor v2 — Python backend with dynamic source engine"
requires-python = ">=3.13"
dependencies = [
    # Web framework
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "python-multipart>=0.0.18",

    # HTTP client
    "httpx>=0.28",

    # Data validation
    "pydantic>=2.10",
    "pydantic-settings>=2.7",

    # Database
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "aiosqlite>=0.20",
    "alembic>=1.14",

    # Auth
    "pyjwt>=2.10",
    "bcrypt>=4.2",

    # Source engine — parsing
    "jsonpath-ng>=1.7",
    "lxml>=5.3",
    "feedparser>=6.0",

    # Source engine — LLM detection (Vertex AI)
    "google-auth>=2.37",
    "requests>=2.32",  # required by google-auth transport

    # Infrastructure
    "redis>=5.2",
    "cryptography>=44.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "httpx",  # for TestClient
    "ruff>=0.8",
]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
